refactor(petcore): single RPC method registry, typed errors and pause reasons
- rpc/methods.rs is now the only place that declares the 56-method
  inventory; discovery, param allowlists, and dispatch all read it
- PetCoreError gains InvalidParams and GenerationConflict variants;
  wire codes/messages stay byte-identical (parity tests added), legacy
  string sniffing kept as fallback for unmigrated producers
- app_server exports ExternalCheckpointPauseKind so generation maps
  recoverable pauses from a typed value instead of English substrings
- generation.rs splits into generation/{mod,visual_qa,messages}; the
  visual-production QA subsystem moves to visual_qa.rs, duplicated
  persisted copy moves to messages.rs
- DEFAULT_PET_STATES_JSON exists once in storage/mod.rs and the
  migration interpolates it
- fs_identity::same_file centralizes the stat-sandwich dev/ino checks
